feat(home): add featured courses section with course cards and view all button
